Move DDL methods out of AdapterBackedSqlClient
The DDL methods in AdapterBackedSqlClient come from the period when
they were part of the core MetricFlow SqlClient interface. Now that
there is an extended protocol in the test packages, these methods can
be moved out of the production AdapterBackedSqlClient and into a
test package extension.

This provides us with a bit more flexibility to take temporary short cuts
in order to quickly enable support for new adapter types without exposing
that silliness in any of our production code.
